Depricate GlobalAssemblyInfo.cs file
Replace CommonAssemblyInfo.cs with the GlobalAssemblyInfo.cs file auto-generated by the AssemblyInfo task in the Build.proj file.
Add GlobalAssemblyInfo.cs file to .gitignore since its generated and does not need to be in source control.
The only side-effect is that when someone initially pulls down the code, they will need to run \build.cmd before they can compile in VS, since it will be looking for the generated file.