Introduce GlobalAssemblyInfo.proj input file
Modify Build.proj to import an optional GlobalAssemblyInfo.proj file, which provides all inputs for the AssemblyInfo task.
This provides a clean separation of the build logic from the build input.  This would potentially be valuable if I were to suggest some of my changes to the MSBuildTasks project (https://github.com/loresoft/msbuildtasks) in a pull request.