Remove build configuration from AssemblyInfo
Since /build.cmd needs to be run once to gen the GlobalAssemblyInfo.cs file so the VS compile can work (which is in Debug configuration mode), it doesn't make sense to embed a configuration mode of Release into the file.
So removing it for now until I can come up with a better way to integrate the two together.