Create new Version.targets file with moved tasks
Move the Version target from Build.proj to the new Version.targets file with associated properties.
Move call to execute the Version target to the CallTarget task in the Build target instead of using DependsOnTargets in the Compile target.
Also fix Attrib task so that it uses SourceRootPath property instead of MSBuildProjectDirectory since without that its not actually working.