Convert Package.targets to use nuget spec .csproj
The previous approach was to completely gen the .nuspec file and feed it into 'nuget spec'.  However, that produces a lot of redundancy between what's already configured in the source code .csproj file.  It's easier to use the .csproj file as the source and use the .nuget file for overrides or additional metadata.
To make Package.targets more parameterized, created the following properties in Input.targets:
- Input_PackageProjectDirPath
- Input_PackageProjectName
- Input_PackageTemplateSpecFilePath