Modify Version target to gen multiple AssemblyInfo's
The original design was for the Version target to generate a single GlobalAssemblyInfo.cs that was linked into all projects in the solution.  The problem was it left out a few key assembly attributes that were still being defined manually in each project's AssemblyInfo.cs file, namely AssemblyTitle and AssemblyDescription.
This change feeds all of that metadata into the Version target via an item group in Input.targets (Input_ProjectDirs).  Once there, the Version target generates individual AssemblyInfo files within each project directory that contain common and specific metadata, all sourced from the Input.targets file.
The final important connection is that a lot of this metadata is then read by NuGet to create the package.  Now you have one place to source the metadata that feeds the AssemblyInfo files, which in turn get written to the assemblies themselves and the Nuget package.

Also improve XML comments in both the Input.targets and Version.targets files.